Simplify our error trapping

+++ lang/shell/bash/include/header.sh
@@ -26,7 +26,7 @@ __koopa_bash_source_dir() { # {{{1
     return 0
 }
 
-__koopa_cleanup() {
+__koopa_exit_trap() {
     # """
     # Kill all processes whose parent is this process.
     # @note Updated 2022-04-11.
@@ -35,13 +35,22 @@ __koopa_cleanup() {
     # - https://linuxize.com/post/kill-command-in-linux/
     # - https://stackoverflow.com/questions/28657676/
     # - https://stackoverflow.com/questions/41370092/
+    # - https://stackoverflow.com/questions/65420781/
     # - https://unix.stackexchange.com/questions/222307/
     # - https://unix.stackexchange.com/questions/240723/
     # - https://unix.stackexchange.com/questions/256873/
     # - https://unix.stackexchange.com/questions/478281/
     # - https://www.networkworld.com/article/3174440/
     # """
-    pkill -P "${$}"
+    if [[ "${?}" -gt 0 ]]
+    then
+        if __koopa_is_installed 'ps'
+        then
+            ps -p "${$}"
+        fi
+        pkill -P "${$}"
+    fi
+    return 1
 }
 
 __koopa_is_installed() { # {{{1
@@ -167,14 +176,7 @@ __koopa_bash_header() { # {{{1
     fi
     if [[ "${dict[activate]}" -eq 0 ]]
     then
-        for sig in INT QUIT HUP TERM
-        do
-            trap "
-                __koopa_cleanup
-                trap - ${sig} EXIT
-                kill -s ${sig} "'"$$"' "${sig}"
-        done
-        trap __koopa_cleanup EXIT
+        trap __koopa_exit_trap EXIT
         # Fix for RHEL/CentOS/Rocky Linux 'BASHRCSOURCED' unbound variable.
         # https://100things.wzzrd.com/2018/07/11/
         #   The-confusing-Bash-configuration-files.html
